Feat: 세로모드 잠금, 포켓몬 이미지 크기 변경
포켓몬 지역별로 불러오기 메서드 분리, 샤이니 포켓몬 뷰모델 생성, 포켓몬 디테일 뷰모델 생성, 다크모드 테마 뷰모델 생성
