+++ lib/utils/string_utils.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:pokedex_app/models/pokedex_entry.dart';
 import 'package:pokedex_app/models/pokemon_detail.dart';
 
 final pokemonTypeColors = {
@@ -75,25 +74,6 @@ extension StringCapitalize on String {
         ],
       );
     }).toList();
-  }
-}
-
-class ShinyPokemon {
-  bool isShiny = false; // 이로치 상태 기본값 설정
-
-  //이로치 상태 토글하는 메서드 생성
-  void toggleShiny() {
-    isShiny = !isShiny;
-  }
-
-  //포켓몬id로 이미지 url을 반환
-  String getPokemonImageUrl(PokedexEntry url) {
-    final pokemonId = url.getPokemonId();
-    if (isShiny) {
-      return 'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/shiny/$pokemonId.png';
-    } else {
-      return 'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/$pokemonId.png';
-    }
   }
 }
