+++ lib/viewmodels/theme_viewmodel.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class ThemeViewModel with ChangeNotifier {
+  bool _isDarkMode = false;
+
+  bool get isDarkMode => _isDarkMode;
+
+  void toggleTheme(bool value) {
+    _isDarkMode = value;
+    notifyListeners();
+  }
+}
