+++ pubspec.yaml
@@ -37,6 +37,7 @@ dependencies:
   http: ^1.3.0
   cached_network_image: ^3.3.0
   path_provider: ^2.1.1
+  provider: ^6.1.2
 
 dev_dependencies:
   flutter_test:
